# resources/design/openapi.yaml
openapi: "3.1.0"
info:
  title: AgroIA API
  description: API de AgroInteligente Colombia — Plataforma de diagnóstico agronómico inteligente
  version: "0.2.0"
  contact:
    name: AgroIA Team

servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
  - url: https://api.agroia.co/api/v1
    description: Producción

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT RS256 emitido por el Auth Service

  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Código de error único
          example: VALIDATION_ERROR
        message:
          type: string
          description: Mensaje descriptivo en español
          example: El campo 'email' es requerido

    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8

  responses:
    "400":
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "401":
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "403":
      description: Acceso denegado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "404":
      description: No encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "409":
      description: Conflicto
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "500":
      description: Error interno
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

paths:
  /recomendaciones/analyze:
    post:
      summary: Analizar aptitud del suelo
      description: Pipeline completo — datos de suelo → ML → reglas → discordancia → respuesta
      tags: [Recomendaciones]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [finca_id]
              properties:
                finca_id:
                  type: string
                  format: uuid
                  description: UUID de la finca a analizar
                cultivo_id:
                  type: string
                  format: uuid
                  description: UUID del cultivo objetivo (opcional)
      responses:
        "200":
          description: Recomendación generada exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendResponse"
        "422":
          $ref: "#/components/responses/422"

  /recomendaciones/historial/{finca_id}:
    get:
      summary: Historial de recomendaciones por finca
      tags: [Recomendaciones]
      security:
        - bearerAuth: []
      parameters:
        - name: finca_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Lista paginada de recomendaciones

  /costeo/parametros/vigentes:
    get:
      summary: Conjunto de parámetros de costeo publicado y vigente
      description: Devuelve el conjunto vigente a la fecha con servicios, componentes, tramos, factores, política, impuestos, descuentos, densidad y zonas. Roles Admin y Agrónomo.
      tags: [Costeo]
      security:
        - bearerAuth: []
      parameters:
        - name: fecha
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Fecha de referencia (hoy por defecto)
      responses:
        "200":
          description: Conjunto vigente serializado
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"

  /costeo/simular:
    post:
      summary: Calcular costeo sin persistir (motor determinista)
      description: Ejecuta la tubería de cálculo AGC-COST sobre el conjunto indicado o el vigente. Roles Admin y Agrónomo.
      tags: [Costeo]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [area_ha, servicios]
              properties:
                conjunto_id:
                  type: string
                  format: uuid
                fecha_referencia:
                  type: string
                  format: date
                area_ha:
                  type: number
                puntos:
                  type: integer
                km:
                  type: number
                dificultad:
                  type: string
                departamento:
                  type: string
                municipio:
                  type: string
                descuento_pct:
                  type: number
                servicios:
                  type: array
                  items:
                    type: object
                    properties:
                      codigo:
                        type: string
                      cantidad:
                        type: number
      responses:
        "200":
          description: Desglose completo (líneas, factores, totales, advertencias, indicadores y traza)
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
        "422":
          description: Error de negocio del motor (TRAMOS_INCONSISTENTES, DESCUENTO_EXCEDE_TOPE, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"